Conflicts:
	ThinKB/src/main/java/com/kb/star/controller/IdeaRoomController.java
	ThinKB/src/main/webapp/WEB-INF/views/firstMeeting/meetingList.jsp
	ThinKB/src/main/webapp/WEB-INF/views/firstMeeting/roomStage1.jsp

+++ ThinKB/src/main/java/com/kb/star/controller/IdeaOpinionsController.java
@@ -19,10 +19,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import com.kb.star.command.addFunction.AddCommand;
 import com.kb.star.command.room.IdeaOpinions2Command;
+import com.kb.star.command.room.IdeaOpinionsClear2Command;
 import com.kb.star.command.room.IdeaOpinionsClearCommand;
 import com.kb.star.command.room.IdeaOpinionsCommand;
 import com.kb.star.dto.IdeaOpinionsDto;
+import com.kb.star.dto.Ideas;
+import com.kb.star.dto.MeetingRooms;
 import com.kb.star.util.IdeaOpinionsDao;
+import com.kb.star.util.RoomDao;
 
 
 @Controller
@@ -163,6 +167,11 @@ public class IdeaOpinionsController {
     	model.addAttribute("request", request);
     	model.addAttribute("roomId", roomId);
         model.addAttribute("ideaId", ideaId);
+        
+        // 방장 사이드탭
+        RoomDao dao=sqlSession.getMapper(RoomDao.class);
+        MeetingRooms info = dao.roomDetailInfo(roomId);
+        model.addAttribute("meetingRoom", info);
     	
         return "/firstMeeting/ideaOpinionsClear";
     }
@@ -238,17 +247,9 @@ public class IdeaOpinionsController {
                 hatColor = "Smart";
         }
         opinionForm.setHatColor(hatColor);  // HatColor 값을 설정합니다.
+        
 
         IdeaOpinionsDao ideaOpinionsDao = sqlSession.getMapper(IdeaOpinionsDao.class);
-        
-        // 사용자가 이미 이 탭에 의견을 작성했는지 확인
-        int existingOpinions = ideaOpinionsDao.countUserOpinionsInTab(userId, ideaId, hatColor, 2);
-        int existingOpinions2 = ideaOpinionsDao.countUserOpinionsInTab(userId, ideaId, currentTab, 2);
-        if (existingOpinions2 > 0) {
-            model.addAttribute("alreadyWritten", true);
-        }
-
-        
         ideaOpinionsDao.insertOpinion2(opinionForm);
         
         return "redirect:/ideaOpinions2?currentTab=" + currentTab + "&roomId=" + roomId + "&ideaId=" + ideaId;
@@ -279,13 +280,20 @@ public class IdeaOpinionsController {
 
         IdeaOpinionsDao ideaOpinionsDao = sqlSession.getMapper(IdeaOpinionsDao.class);
 
+        // 좋아요 수 가져오기
+        // int likeNum = ideaOpinionsDao.getLikeNum(opinionId);
+        // 사용자가 이미 좋아요를 눌렀는지 확인
+        boolean alreadyLiked = checkIfUserLikedOpinion(userId, opinionId);
+        
         // 좋아요 수 증가 또는 감소
-        if (like) {
+        if (like && !alreadyLiked) {
             ideaOpinionsDao.increaseLikeNum(opinionId);  // 좋아요 수 증가
             ideaOpinionsDao.addUserLike(userId, opinionId);  // 좋아요 추가
-        } else {
+            ideaOpinionsDao.updateContributionLikeNum(ideaId, userId, roomId, true); // 좋아요 수만큼 기여도 증가
+        } else if (!like && alreadyLiked) {
             ideaOpinionsDao.decreaseLikeNum(opinionId);  // 좋아요 수 감소
             ideaOpinionsDao.removeUserLike(userId, opinionId);  // 좋아요 제거
+            ideaOpinionsDao.updateContributionLikeNum(ideaId, userId, roomId, false); // 좋아요 수만큼 기여도 증가
         }
 
         int updatedLikeNum = ideaOpinionsDao.getLikeNum(opinionId);
@@ -300,7 +308,39 @@ public class IdeaOpinionsController {
     	IdeaOpinionsDao ideaOpinionsDao = sqlSession.getMapper(IdeaOpinionsDao.class);
         return ideaOpinionsDao.checkUserLikedOpinion(userId, opinionId);
     }
+    
+	
+    // ideaOpinionsClear2.jsp
+    @RequestMapping("/ideaOpinionsClear2")
+    public String ideaOpinionsClear2(HttpServletRequest request, Model model,
+						    		@RequestParam("roomId") int roomId, @RequestParam("ideaId") int ideaId) {
+    	model.addAttribute("request", request);
+    	model.addAttribute("roomId", roomId);
+        model.addAttribute("ideaId", ideaId);
+        
+        IdeaOpinionsClear2Command ideaOpinionsClear2Command = new IdeaOpinionsClear2Command(sqlSession);
+		ideaOpinionsClear2Command.execute(model);
+        
+        // 방장 사이드탭
+        RoomDao dao=sqlSession.getMapper(RoomDao.class);
+        MeetingRooms info = dao.roomDetailInfo(roomId);
+        model.addAttribute("meetingRoom", info);
+        
+        // Ideas 테이블에서 Title과 StageID 가져오기
+//        List<Ideas> ideasInfo = ideaOpinionsDao.getIdeasInfo(roomId);
+//        model.addAttribute("ideasInfo", ideasInfo);
+    	
+		return "firstMeeting/ideaOpinionsClear2";
+    }
 
+	// stage 5로 이동 = IdeaRoomController의 case 5 = (방장)보고서 작성화면/(사용자)요약보고서
+	@RequestMapping("/goStage5")
+	public String goStage5(@RequestParam("roomId") int roomId,
+		            	   HttpServletRequest request, Model model) {
+		model.addAttribute("request", request);
+		model.addAttribute("roomId", roomId);
+		return "redirect:/roomDetail";
+	}
    
 }
